--- efficient/__init__.py
@@ -1,3 +1,0 @@
-def dummy_function():
-
-    print("Hello World!")
