--- efficient/custom_attacker.py
@@ -1,5 +1,0 @@
-from allennlp.interpret.attackers.attacker import Attacker
-
-
-class DummyAttacker(Attacker):
-    pass
